requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Mail>=0.9.1
Flask-Login>=0.6.3
Flask-JWT-Extended>=4.6.0
Flask-WTF>=1.2.1
Flask-Limiter>=3.5.0
pandas>=2.2.0
numpy>=1.26.3
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0
xgboost>=2.0.3
shap>=0.44.1
matplotlib>=3.8.2
seaborn>=0.13.1
scipy>=1.12.0
requests>=2.31.0
pytest>=8.0.0
beautifulsoup4>=4.12.3
openpyxl>=3.1.2
bcrypt>=4.1.2
xhtml2pdf>=0.2.11
python-dotenv>=1.0.1
gunicorn>=21.2.0
Werkzeug>=3.0.1
itsdangerous>=2.1.2
psycopg2-binary>=2.9.9
